# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TapasHub Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/company-summary/{companyId}:
    get:
      operationId: getCompanySummary
      tags: [dashboard]
      summary: KPI summary for a specific company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company KPI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Revenue chart data (last 12 months)
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Revenue series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed across all companies
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── Orders ──────────────────────────────────────────────────────
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{employeeId}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/attendance-summary:
    get:
      operationId: getAttendanceSummary
      tags: [employees]
      summary: Attendance summary (present/absent/leave today)
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # ─── Customers ───────────────────────────────────────────────────
...
components:
  schemas:
...
    CompanySummary:
      type: object
      required: [companyId, companyName, revenue, orders, employees, profit, growth]
      properties:
        companyId:
          type: integer
        companyName:
          type: string
        companySlug:
          type: string
        revenue:
          type: number
        orders:
          type: integer
        employees:
          type: integer
        profit:
          type: number
        growth:
          type: number
        inventoryValue:
          type: number
        openLeads:
          type: integer
    ChartDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        secondary:
          type: ["number", "null"]
    ActivityItem:
      type: object
      required: [id, type, title, description, companyName, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [order, payment, lead, employee, inventory, approval, customer]
        title:
          type: string
        description:
          type: string
        companyName:
          type: string
        timestamp:
          type: string
        amount:
          type: ["number", "null"]
        status:
          type: ["string", "null"]

    # ─── Orders ────────────────────────────────────────────────────
...
    Employee:
      type: object
      required: [id, companyId, companyName, firstName, lastName, email, department, designation, employeeCode, status, joinDate, salary, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        designation:
          type: string
        employeeCode:
          type: string
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        joinDate:
          type: string
        salary:
          type: number
        managerId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [companyId, firstName, lastName, email, department, designation, salary, joinDate]
      properties:
        companyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        designation:
          type: string
        salary:
          type: number
        joinDate:
          type: string
        managerId:
          type: integer
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        department:
          type: string
        designation:
          type: string
        salary:
          type: number
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        managerId:
          type: integer
    EmployeeList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AttendanceSummary:
      type: object
      required: [date, totalEmployees, present, absent, onLeave, late]
      properties:
        date:
          type: string
        totalEmployees:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        onLeave:
          type: integer
        late:
          type: integer

    # ─── Customers ─────────────────────────────────────────────────
...